requirements.txt
requests
dnspython
colorama
scapy
python-whois